Correcting the calculation of the frequency of a compound in a group.
It was calculating the inverse. Instead of using the threshold as the minimum frequency a compound should have within a group (proportion of samples in which the compound is present), it was using it as the minimum frequency of absence of a compound in a group (proportion of samples in which the compound is absent).