Correct bug in kovats_retention_index.R
There was a bug with the function, some of the last compounds were getting a RI lower than that of the alkane before them.
This was happening because the correction for the calculation of the RI for peaks with no alkane after them was adding the RT of the target peak to the RI of the last compound for which a RI was calculated. Therefore, in cases where the last compound for which the RI was calculated had a RT lower than the alkane before the target compound (i.e. the alkane is not in the samples, but in the standards), the RT of the target peak was being summed to the RI of a peak that eluded before the alkane that would precede the target peak. 

Now, if there is no alkane after the target peak, the RI of the target peak is calculated as the sum between the RI of previous the alkane and the RT of the target peak.