New version of correct alignment (not yet complete)
This new version will do both the addition of new empty peaks (if needed) and the displacement of peaks. the seaparate usage of add_empty_peaks will no longer be needed.
The idea is to also imporve the peak values displacemetn operation to avoid step-by-step movements, directly indicating the original and final peak for a value within a sample. Move_one_peak separate fucntion will no longer be needed.
Once the new version is complete and it has been tested for correct functioning, the old fucntion versions will be removed, and the new version will be renamed to just correct_alignment.