new function shape_aligned_table
Is a new version of shape_group_table that integrates the filtering of trace compounds (trace_comps), removing (or not) of unidentified compounds (drop_na_compounds), and the calculation of the RI (kovats_retention_index ).
The function kovats_retention_index remains as an independent function (the other integrated functions were removed). Its code is long and it will be easier to modify (if needed) outside of another fucntion, and it also could be used in isolation, outside of this new function.